+++ elink.jl
@@ -1,24 +1,24 @@
 
-function mean(a::Float64)
+function mean(a::Float)
     a
 end
 
 wd=(ARGS[1])
 cd(wd)
 
-S = dlmread("class_ids.csv",",",Float64)
+S = dlmread("class_ids.csv",",",Float)
 S=int(S[2:end,2:end])
 
     n=size(S,1)
     Z=zeros(Any,n-1,3)
-    R=zeros(Float64,n+n-1,n-1)
+    R=zeros(Float,n+n-1,n-1)
     R[1:n,1]=1:n # R indexes leafs and groups
     MID=1:n # MID indexes current groups
     
     
     maxpair=zeros(Any,3,1)
     
-    meanDI=zeros(Float64,n,n)
+    meanDI=zeros(Float,n,n)
     for i=1:(n-1)
         for j=(i+1):n
             meanDI[i,j]=mean(int(S[j,:].==S[i,:]))
@@ -48,8 +48,8 @@ S=int(S[2:end,2:end])
         an2=int(R[int(Z[s-1,2]),find(R[int(Z[s-1,2]),:].!=0)])
         allnode = [an1 an2] #just merged nodes
         MID = [MID, n+(s-1)] #new group
-        meanDI=[meanDI zeros(Float64,(n-s),1)]  # add a column of zeros
-        meanDI=[meanDI;zeros(Float64,1,(n-s)+1)]# add a row of zeros for new group
+        meanDI=[meanDI zeros(Float,(n-s),1)]  # add a column of zeros
+        meanDI=[meanDI;zeros(Float,1,(n-s)+1)]# add a row of zeros for new group
         
         DIC=ones(n-s,1)
